Unpin Ansible Version
By default [Vagrant installs the latest from EPEL][vagrant-docs],
which should be more than sufficient.

This was causing an error because version 2.1.2.0 was not yet
published to EPEL but was present in pip. The pip installation
method is seemingly fraught with many other pitfalls, so we will
avoid doing that.

 [vagrant-docs]: https://j.mp/2e8ZHlt